requirements.txt
streamlit
plotly
pytz
streamlit-option-menu
streamlit
streamlit-option-menu
pandas
matplotlib
seaborn
scikit-learn
pillow
openpyxl
